Add Domain
Two structs are added, Domain and DomainDim. The purpose of both is to
describe the layout of arrays representing attributes of the fluid, and
spatial properties of the simulation domain.

DomainDim stores the number (n) of elements along dim, the stride
between elements along dim, the spatial distance (dx) between elements.

Domain stores the total number of elements (N) across all dims, and
a DomainDim member for each dimension.