Rename 'p' -> 'f' and 'div' -> 'g' to match CG
In Jacobi and Gauss-Seidel solve functions, the input arguments 'p' and
'div' are renamed to 'f' and 'g', to match the CG/PCG functions.

The reason for choosing 'f' and 'g' over 'p' and 'div', is that 'p' is
already used in CG/PCG for the search vector, a common convention.
Additionally, these functions have no requirement that 'f' is the
pressure, and 'g' the divergence of velocity field.